Rename AlgorithmConfig to DigitizationConfig
