+++ Examples/Algorithms/Digitization/include/ActsExamples/Digitization/DigitizationConfig.hpp
@@ -9,13 +9,21 @@
 #pragma once
 
 #include "Acts/Definitions/Algebra.hpp"
+#include "Acts/Geometry/GeometryHierarchyMap.hpp"
+#include "Acts/Geometry/TrackingGeometry.hpp"
 #include "Acts/Utilities/BinUtility.hpp"
 #include "Acts/Utilities/BinningType.hpp"
+#include "Acts/Utilities/Logger.hpp"
+#include "ActsExamples/Digitization/DigitizationConfig.hpp"
 #include "ActsExamples/Digitization/SmearingConfig.hpp"
+#include "ActsExamples/Framework/IAlgorithm.hpp"
 #include "ActsExamples/Framework/RandomNumbers.hpp"
+#include "ActsExamples/Utilities/OptionsFwd.hpp"
 #include "ActsFatras/Digitization/UncorrelatedHitSmearer.hpp"
 
 #include <functional>
+#include <memory>
+#include <string>
 
 namespace ActsExamples {
 
@@ -39,7 +47,7 @@ using VarianceGenerator =
 /// The BinUtility defines the segmentation and which parameters
 /// are defined by this.
 ///
-struct GeometricDigitizationConfig {
+struct GeometricConfig {
   std::vector<Acts::BoundIndices> indices = {};
   Acts::BinUtility segmentation;
   /// Drift generation
@@ -67,9 +75,51 @@ struct GeometricDigitizationConfig {
 /// It contains:
 /// - optional GeometricConfig
 /// - optional SmearingConfig
-struct DigitizationConfig {
-  GeometricDigitizationConfig geometricDigiConfig;
+struct DigiComponentsConfig {
+  GeometricConfig geometricDigiConfig;
   SmearingConfig smearingDigiConfig = {};
 };
+
+class DigitizationConfig {
+ public:
+  DigitizationConfig(const Options::Variables &vars)
+      : DigitizationConfig(vars,
+                        Acts::GeometryHierarchyMap<DigiComponentsConfig>()){};
+
+  DigitizationConfig(const Options::Variables &vars,
+                  Acts::GeometryHierarchyMap<DigiComponentsConfig> &&digiCfgs);
+
+  /// Input collection of simulated hits.
+  std::string inputSimHits = "simhits";
+  /// Output source links collection.
+  std::string outputSourceLinks = "sourcelinks";
+  /// Output measurements collection.
+  std::string outputMeasurements = "measurements";
+  /// Output cluster collection.
+  std::string outputClusters = "clusters";
+  /// Output collection to map measured hits to contributing particles.
+  std::string outputMeasurementParticlesMap = "measurement_particles_map";
+  /// Output collection to map measured hits to simulated hits.
+  std::string outputMeasurementSimHitsMap = "measurement_simhits_map";
+  /// Tracking geometry required to access global-to-local transforms.
+  std::shared_ptr<const Acts::TrackingGeometry> trackingGeometry = nullptr;
+  /// Random numbers tool.
+  std::shared_ptr<const RandomNumbers> randomNumbers = nullptr;
+  /// Was the simple smearer requested
+  const bool isSimpleSmearer;
+  /// The digitizers per GeometryIdentifiers
+  Acts::GeometryHierarchyMap<DigiComponentsConfig> digitizationConfigs;
+
+  std::vector<
+    std::pair<Acts::GeometryIdentifier, std::vector<Acts::BoundIndices>>>
+  getBoundIndices();
+
+ private:
+  // Private initializer for SmearingAlgorithm
+  void smearingConfig(const Options::Variables &vars);
+};
+
+std::shared_ptr<ActsExamples::IAlgorithm> createDigitizationAlgorithm(
+  DigitizationConfig &cfg, Acts::Logging::Level lvl);
 
 }  // namespace ActsExamples
