+++ Examples/Algorithms/Digitization/src/DigitizationOptions.cpp
@@ -61,195 +61,3 @@ void ActsExamples::Options::addDigitizationOptions(Description& desc) {
       "values.");
 }
 
-namespace {
-
-enum SmearingTypes : int {
-  eGauss = 0,
-  eGaussTruncated = 1,
-  eGaussClipped = 2,
-  eUniform = 3,
-  eDigital = 4,
-};
-
-constexpr size_t numConfigParametersForType(int type) {
-  // Gaussian smearing requires only a width/standard deviation parameter
-  // Everything else requires a width/pitch and a range
-  return (static_cast<SmearingTypes>(type) == eGauss) ? 1u : 3u;
-}
-
-ActsFatras::SingleParameterSmearFunction<ActsExamples::RandomEngine>
-makeSmearFunctionForType(SmearingTypes smearingType, const double* parameters) {
-  using namespace ActsExamples::Digitization;
-
-  switch (smearingType) {
-    case eGauss:
-      return Gauss(parameters[0]);
-    case eGaussTruncated:
-      return GaussTrunc(parameters[0], {parameters[1u], parameters[2u]});
-    case eGaussClipped:
-      return GaussClipped(parameters[0], {parameters[1u], parameters[2u]});
-    case eUniform:
-      return Uniform(parameters[0], {parameters[1u], parameters[2u]});
-    case eDigital:
-      return Digital(parameters[0], {parameters[1u], parameters[2u]});
-  }
-  return nullptr;
-}
-
-}  // namespace
-
-ActsExamples::Digitization::AlgorithmConfig::AlgorithmConfig(
-    const Options::Variables& vars,
-    Acts::GeometryHierarchyMap<DigitizationConfig>&& digiCfgs)
-    : isSimpleSmearer(vars["digi-smear"].as<bool>()) {
-  digitizationConfigs = std::move(digiCfgs);
-  if (isSimpleSmearer)
-    smearingConfig(vars);
-}
-
-void ActsExamples::Digitization::AlgorithmConfig::smearingConfig(
-    const Options::Variables& variables) {
-  ACTS_LOCAL_LOGGER(
-      Acts::getDefaultLogger("SmearingOptions", Acts::Logging::INFO));
-
-  using namespace Acts::UnitLiterals;
-  using namespace ActsExamples::Options;
-
-  // Smear configuration with command line input
-  // only limited smearing configuration possible:
-  // complex smearing option configuration has to be done with a
-  // digitization config file
-
-  // in case of an error, we always return a configuration struct with empty
-  // smearers configuration. this will be caught later on during the algorithm
-  // construction.
-
-  auto volumes = variables["digi-smear-volume"].as<std::vector<int>>();
-  if (volumes.empty()) {
-    // no configured volumes are not considered an error at this stage
-    return;
-  }
-
-  if (not variables["digi-config-file"].as<std::string>().empty()) {
-    ACTS_WARNING(
-        "Smearing configuration on command-line will override .json "
-        "configuration!");
-  }
-
-  auto indices =
-      variables["digi-smear-indices"].as<std::vector<VariableIntegers>>();
-  auto types =
-      variables["digi-smear-types"].as<std::vector<VariableIntegers>>();
-  auto parameters =
-      variables["digi-smear-parameters"].as<std::vector<VariableReals>>();
-  if (indices.size() != volumes.size()) {
-    ACTS_ERROR("Inconsistent digi-smear-indices options. Expected "
-               << volumes.size() << ", but received " << indices.size());
-    return;
-  }
-  if (types.size() != volumes.size()) {
-    ACTS_ERROR("Inconsistent digi-smear-types options. Expected "
-               << volumes.size() << ", but received " << types.size());
-    return;
-  }
-  if (parameters.size() != volumes.size()) {
-    ACTS_ERROR("Inconsistent digi-smear-parameters options. Expected "
-               << volumes.size() << ", but received " << parameters.size());
-    return;
-  }
-
-  // construct the input for the smearer configuation
-  std::vector<std::pair<Acts::GeometryIdentifier, DigitizationConfig>>
-      smearersInput;
-  for (size_t ivol = 0; ivol < volumes.size(); ++ivol) {
-    Acts::GeometryIdentifier geoId =
-        Acts::GeometryIdentifier(0).setVolume(volumes[ivol]);
-    const auto& volIndices = indices[ivol].values;
-    const auto& volTypes = types[ivol].values;
-    const auto& volParameters = parameters[ivol].values;
-
-    if (volTypes.size() != volIndices.size()) {
-      ACTS_ERROR("Inconsistent number of digi-smear-types values for volume "
-                 << volumes[ivol] << ". Expected " << indices.size()
-                 << ", but received " << types.size());
-      return;
-    }
-    // count the expected number of smearing configuration parameters
-    size_t expectedNumParameters = 0;
-    for (auto smearingType : volTypes) {
-      expectedNumParameters += numConfigParametersForType(smearingType);
-    }
-    if (volParameters.size() != expectedNumParameters) {
-      ACTS_ERROR(
-          "Inconsistent number of digi-smear-parameters values for volume "
-          << volumes[ivol] << ". Expected " << expectedNumParameters
-          << ", but received " << types.size());
-      return;
-    }
-
-    // create the smearing configuration for this geometry identifier
-    SmearingConfig geoCfg;
-    geoCfg.reserve(volIndices.size());
-
-    for (size_t iidx = 0, ipar = 0; iidx < volIndices.size(); ++iidx) {
-      const auto paramIndex = static_cast<Acts::BoundIndices>(volIndices[iidx]);
-      const auto smearingType = static_cast<SmearingTypes>(volTypes[iidx]);
-      const double* smearingParameters = &volParameters[ipar];
-      ipar += numConfigParametersForType(smearingType);
-
-      ParameterSmearingConfig parCfg;
-      parCfg.index = paramIndex;
-      parCfg.smearFunction =
-          makeSmearFunctionForType(smearingType, smearingParameters);
-      if (not parCfg.smearFunction) {
-        ACTS_ERROR("Invalid smearing type for entry "
-                   << iidx << " for volume " << volumes[ivol] << ". Type "
-                   << volTypes[iidx] << " is not valid");
-        return;
-      }
-      geoCfg.emplace_back(std::move(parCfg));
-    }
-    DigitizationConfig dcfg;
-    dcfg.smearingDigiConfig = geoCfg;
-    smearersInput.emplace_back(geoId, std::move(dcfg));
-  }
-  // set the smearer configuration from the prepared input
-  digitizationConfigs =
-      Acts::GeometryHierarchyMap<DigitizationConfig>(std::move(smearersInput));
-}
-
-std::shared_ptr<ActsExamples::IAlgorithm>
-ActsExamples::Options::createDigitizationAlgorithm(
-    ActsExamples::Digitization::AlgorithmConfig& cfg,
-    Acts::Logging::Level lvl) {
-  if (cfg.isSimpleSmearer)
-    return std::make_shared<SmearingAlgorithm>(cfg, lvl);
-  else
-    return std::make_shared<DigitizationAlgorithm>(cfg, lvl);
-}
-
-std::vector<
-    std::pair<Acts::GeometryIdentifier, std::vector<Acts::BoundIndices>>>
-ActsExamples::Options::getBoundIndices(
-    ActsExamples::Digitization::AlgorithmConfig& cfg) {
-  std::vector<
-      std::pair<Acts::GeometryIdentifier, std::vector<Acts::BoundIndices>>>
-      bIndexInput;
-
-  for (size_t ibi = 0; ibi < cfg.digitizationConfigs.size(); ++ibi) {
-    Acts::GeometryIdentifier geoID = cfg.digitizationConfigs.idAt(ibi);
-    const auto dCfg = cfg.digitizationConfigs.valueAt(ibi);
-    std::vector<Acts::BoundIndices> boundIndices;
-    if (cfg.isSimpleSmearer) {
-      for (const auto& sConfig : dCfg.smearingDigiConfig) {
-        boundIndices.push_back(sConfig.index);
-      }
-    } else {
-      boundIndices.insert(boundIndices.end(),
-                          dCfg.geometricDigiConfig.indices.begin(),
-                          dCfg.geometricDigiConfig.indices.end());
-    }
-    bIndexInput.push_back({geoID, boundIndices});
-  }
-  return bIndexInput;
-}
