+++ Examples/Io/Json/include/ActsExamples/Io/Json/JsonDigitizationConfig.hpp
@@ -19,22 +19,22 @@ void to_json(nlohmann::json& j, const ParameterSmearingConfig& psc);
 
 void from_json(const nlohmann::json& j, ParameterSmearingConfig& psc);
 
-void to_json(nlohmann::json& j, const GeometricDigitizationConfig& gdc);
+void to_json(nlohmann::json& j, const GeometricConfig& gdc);
 
-void from_json(const nlohmann::json& j, GeometricDigitizationConfig& gdc);
+void from_json(const nlohmann::json& j, GeometricConfig& gdc);
 
 void to_json(nlohmann::json& j, const SmearingConfig& sdc);
 
 void from_json(const nlohmann::json& j, SmearingConfig& sdc);
 
-void to_json(nlohmann::json& j, const DigitizationConfig& dc);
+void to_json(nlohmann::json& j, const DigiComponentsConfig& dc);
 
-void from_json(const nlohmann::json& j, DigitizationConfig& dc);
+void from_json(const nlohmann::json& j, DigiComponentsConfig& dc);
 
-Acts::GeometryHierarchyMap<DigitizationConfig> readDigiConfigFromJson(const std::string &path);
+Acts::GeometryHierarchyMap<DigiComponentsConfig> readDigiConfigFromJson(const std::string &path);
 
-using DigiConfigContainer = Acts::GeometryHierarchyMap<DigitizationConfig>;
+using DigiConfigContainer = Acts::GeometryHierarchyMap<DigiComponentsConfig>;
 using DigiConfigConverter =
-    Acts::GeometryHierarchyMapJsonConverter<DigitizationConfig>;
+    Acts::GeometryHierarchyMapJsonConverter<DigiComponentsConfig>;
 
 }  // namespace ActsExamples
