+++ Examples/Io/Json/src/JsonDigitizationConfig.cpp
@@ -89,7 +89,7 @@ void ActsExamples::from_json(const nlohmann::json& j,
 }
 
 void ActsExamples::to_json(
-    nlohmann::json& j, const ActsExamples::GeometricDigitizationConfig& gdc) {
+    nlohmann::json& j, const ActsExamples::GeometricConfig& gdc) {
   std::vector<size_t> indices;
   for (const auto& idx : gdc.indices) {
     indices.push_back(static_cast<size_t>(idx));
@@ -102,7 +102,7 @@ void ActsExamples::to_json(
 }
 
 void ActsExamples::from_json(const nlohmann::json& j,
-                             ActsExamples::GeometricDigitizationConfig& gdc) {
+                             ActsExamples::GeometricConfig& gdc) {
   for (const auto& jidx : j["indices"]) {
     gdc.indices.push_back(static_cast<Acts::BoundIndices>(jidx));
   }
@@ -129,7 +129,7 @@ void ActsExamples::from_json(const nlohmann::json& j,
 }
 
 void ActsExamples::to_json(nlohmann::json& j,
-                           const ActsExamples::DigitizationConfig& dc) {
+                           const ActsExamples::DigiComponentsConfig& dc) {
   if (not dc.geometricDigiConfig.indices.empty()) {
     j["geometric"] = nlohmann::json(dc.geometricDigiConfig);
   }
@@ -139,7 +139,7 @@ void ActsExamples::to_json(nlohmann::json& j,
 }
 
 void ActsExamples::from_json(const nlohmann::json& j,
-                             ActsExamples::DigitizationConfig& dc) {
+                             ActsExamples::DigiComponentsConfig& dc) {
   if (j.find("geometric") != j.end()) {
     nlohmann::json jgdc = j["geometric"];
     from_json(jgdc, dc.geometricDigiConfig);
@@ -151,7 +151,7 @@ void ActsExamples::from_json(const nlohmann::json& j,
 }
 
 
-Acts::GeometryHierarchyMap<ActsExamples::DigitizationConfig>
+Acts::GeometryHierarchyMap<ActsExamples::DigiComponentsConfig>
 ActsExamples::readDigiConfigFromJson(const std::string& path) {
   nlohmann::json djson;
   auto in = std::ifstream(path, std::ifstream::in | std::ifstream::binary);
